feat: cfd-364 generate netex for hopper fares (#104)
* save progress on netex

* add missing imports

* get typescript compiling

* get tests passing

* fix linting issues

* update a few npm packages to fix vulnerabilities

* update npm packages to fix moderate vulnerabilities

* update npm packages to fix moderate vulnerabilities

* format the json
