+++ repos/fdbt-site/src/pages/api/serviceList.ts
@@ -1,9 +1,10 @@
 import { NextApiResponse } from 'next';
 import isArray from 'lodash/isArray';
-import { SelectedService, NextApiRequestWithSession } from '../../interfaces/index';
+import { NextApiRequestWithSession } from '../../interfaces/index';
 import { getFareTypeFromFromAttributes, redirectTo, redirectToError } from './apiUtils';
 import { SERVICE_LIST_ATTRIBUTE } from '../../constants/attributes';
 import { updateSessionAttribute } from '../../utils/sessions';
+import { SelectedService } from 'shared/matchingJsonTypes';
 
 const errorId = 'checkbox-0';
